Enabled single-processor PC calibration

+++ Code/PCCalGUI.m
@@ -272,38 +272,47 @@ if handles.VanPont
 
     CalibrationPointsPC = zeros(npoints,3);
     
-    NumCores = Settings.DoParallel;
-    try
-        ppool = gcp('nocreate');
-        if isempty(ppool)
-            parpool(NumCores);
-        end
-    catch
-        ppool = matlabpool('size');
-        if ~ppool
-            matlabpool('local',NumCores); 
-        end
-    end
-    M = NumCores;
-    
-    pctRunOnAll javaaddpath('java')
-    ppm = ParforProgMon( 'Point Calibration ', npoints,1,400,50 );
-%     profile on
     Algorithm = handles.Algorithm;
     ScanParams.xstar = Settings.XStar(1);
     ScanParams.ystar = Settings.YStar(2);
     ScanParams.zstar = Settings.ZStar(3);
-    parfor (i=1:npoints,M)
-        PCref = PCMinSinglePattern(Settings, ScanParams, Settings.CalibrationPointIndecies(i),Algorithm);
-        disp(['Point: ' num2str(i)])
-        CalibrationPointsPC(i,:) = PCref';
-        ppm.increment();
+    
+    %Perform Calibration
+    if Settings.DoParallel > 1
+        NumCores = Settings.DoParallel;
+        try
+            ppool = gcp('nocreate');
+            if isempty(ppool)
+                parpool(NumCores);
+            end
+        catch
+            ppool = matlabpool('size');
+            if ~ppool
+                matlabpool('local',NumCores); 
+            end
+        end
+        M = NumCores;
+        pctRunOnAll javaaddpath('java')
+        ppm = ParforProgMon( 'Point Calibration ', npoints,1,400,50 );
+    
+        parfor (i=1:npoints)
+            PCref = PCMinSinglePattern(Settings, ScanParams, Settings.CalibrationPointIndecies(i),Algorithm);
+            disp(['Point: ' num2str(i)])
+            CalibrationPointsPC(i,:) = PCref';
+            ppm.increment();
+        end
+        ppm.delete();
+    else
+        %Single Processor Calculation
+        for i = 1:npoints
+            PCref = PCMinSinglePattern(Settings, ScanParams, Settings.CalibrationPointIndecies(i),Algorithm);
+            disp(['Point: ' num2str(i)])
+            CalibrationPointsPC(i,:) = PCref';
+        end
     end
-%     profile off
-%     profile viewer
-    ppm.delete();
+    
     Settings.CalibrationPointsPC = CalibrationPointsPC;
-
+    
     psize = Settings.PhosphorSize;
     
     handles.MeanXstar = mean(Settings.CalibrationPointsPC(:,1)+(Settings.XData(Settings.CalibrationPointIndecies))/psize);
